Refactor: change Double to BigDecimal

+++ src/main/java/by/osinovi/orderservice/dto/message/PaymentMessage.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -17,5 +18,5 @@ public class PaymentMessage {
     private Long orderId;
     private Long userId;
     private PaymentStatus status;
-    private Double paymentAmount;
+    private BigDecimal paymentAmount;
 }
